`DataColumnsAlignWithBlock`: Move into its own file.
